moved memory structure to dedicated header.
